role permissions

+++ app/Http/Controllers/Admin/RoleController.php
@@ -3,11 +3,14 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Admin\StoreRoleRequest;
-use App\Http\Requests\Admin\UpdateRoleRequest;
+use App\Http\Requests\Admin\Role\StoreRoleRequest;
+use App\Http\Requests\Admin\Role\UpdateRoleRequest;
 use App\Permission;
 use App\Role;
 use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Support\Facades\Auth;
+
+;
 
 class RoleController extends Controller{
     /**
@@ -16,7 +19,11 @@ class RoleController extends Controller{
      * @return \Illuminate\Http\Response
      */
     public function list() {
-        return view('themes.admin.html.role.roles', ['roles' => Role::all()]);
+
+        if(Auth::user()->can('viewList', Role::class))
+            return view('themes.admin.html.role.roles', ['roles' => Role::all()]);
+
+        return redirect()->route('admin.no-access');
     }
 
     /**
@@ -25,24 +32,34 @@ class RoleController extends Controller{
      * @return \Illuminate\Http\Response
      */
     public function create() {
-        return view('themes.admin.html.role.new', [
-            'roles' => Role::all(),
-            'permissions' => Permission::all()
-        ]);
+        if(Auth::user()->can('create', Role::class)){
+            return view('themes.admin.html.role.new', [
+                'roles' => Role::all(),
+                'permissions' => Permission::all()
+            ]);
+        }
+        else {
+            return redirect()->route('admin.no-access');
+        }
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \App\Http\Requests\Admin\StoreRoleRequest $request
+     * @param  \App\Http\Requests\Admin\Role\StoreRoleRequest $request
      * @return \Illuminate\Http\Response
      */
     public function store(StoreRoleRequest $request) {
-        $role = $this->save($request);
 
-        $role->permissions()->sync($request->permissions, false);
+        if(Auth::user()->can('create', Role::class)){
+            $role = $this->save($request);
 
-        return redirect()->route('admin.role.show', $role->id);
+            $role->permissions()->sync($request->permissions, false);
+
+            return redirect()->route('admin.role.show', $role->id);
+        }
+        else
+            return redirect()->route('admin.no-access');
     }
 
     /**
@@ -54,11 +71,16 @@ class RoleController extends Controller{
     public function show($id) {
         $role = Role::find($id);
 
-        return view('themes.admin.html.role.role',
-            ['role' => $role,
-             'parent' => $role->parent,
-             'permissions' => $role->permissions
-            ]);
+        if(Auth::user()->can('view', $role)){
+            return view('themes.admin.html.role.role',
+                ['role' => $role,
+                 'parent' => $role->parent,
+                 'permissions' => $role->permissions
+                ]);
+        }
+        else {
+            return redirect()->route('admin.no-access');
+        }
     }
 
     /**
@@ -70,28 +92,41 @@ class RoleController extends Controller{
     public function edit($id) {
         $role = Role::find($id);
 
-        return view('themes.admin.html.role.edit',
-            [
-                'role' => $role,
-                'roles' => Role::whereNotIn('id', array_merge([$id], $role->descendants->pluck('id')->toArray()))->get(),
-                'permissions' => Permission::all(),
-                'role_permissions' => $role->permissions->pluck('id')
-            ]);
+        if(Auth::user()->can('update', $role)){
+            return view('themes.admin.html.role.edit',
+                [
+                    'role' => $role,
+                    'roles' => Role::whereNotIn('id', array_merge([$id], $role->descendants->pluck('id')->toArray()))->get(),
+                    'permissions' => Permission::all(),
+                    'role_permissions' => $role->permissions->pluck('id')
+                ]);
+        }
+        else {
+            return redirect()->route('admin.no-access');
+        }
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \App\Http\Requests\Admin\UpdateRoleRequest $request
+     * @param  \App\Http\Requests\Admin\Role\UpdateRoleRequest $request
      * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function update(UpdateRoleRequest $request, $id) {
-        $role = $this->save($request, $id);
 
-        $role->permissions()->sync($request->permissions);
+        $role = Role::find($id);
 
-        return redirect()->route('admin.role.show', $id);
+        if($role && Auth::user()->can('update', $role)){
+
+            $this->save($request, $role)
+                ->permissions()->sync($request->permissions);
+
+            return redirect()->route('admin.role.show', $id);
+        }
+        else {
+            return redirect()->route('admin.no-access');
+        }
     }
 
     /**
@@ -101,22 +136,28 @@ class RoleController extends Controller{
      * @return \Illuminate\Http\Response
      */
     public function destroy($id) {
-        Role::destroy($id);
 
-        return redirect()->route('admin.roles.list');
+        $role = Role::find($id);
+
+        if($role && Auth::user()->can('delete', $role)){
+            Role::destroy($id);
+
+            return redirect()->route('admin.roles.list');
+        }
+        else {
+            return redirect()->route('admin.no-access');
+        }
     }
 
     /**
      * @param FormRequest $request
-     * @param null $id
+     * @param null $role
      * @return Role
      */
-    private function save(FormRequest $request, $id = null) {
+    private function save(FormRequest $request, $role = null) {
 
-        if(is_null($id))
+        if(is_null($role))
             $role = new Role();
-        else
-            $role = Role::find($id);
 
         $role->name = $request->name;
         $role->parent_id = $request->parent_id;
